Create README.md
Project Overview:
The purpose of this project is to see the relationship between interest rate and GDP in U.S. The macroeconomics theory states that there is a negative relation between interest rate and GDP. In this project, we have time series data of interest rate and GDP rate for the U.S. We will use linear regression analysis to test this theory.
Project Structure:
The project will start by providing a data report. The data report will explain where the data comes from and how to access the data.
The second part of the project will provide summary statistics about real interest rate and GDP trends and provide visualizations of these statistics to help further illustrate the project.
The third part of the project will state a hypothesis and compared with trends in previous scholarship. Then, a linear regression will be performed to test the hypothesis and the results will be stated. Lastly, there will be a detailed discussion of the results and how it follows or goes against previous scholarship.
Data Report:
The data for this project comes from the Federal Reserve Economic Data website (FRED). FRED provides monthly data on the U.S. interest rate, GDP from 1948 onward, and it identifies times of periods of economic contraction. The data can be accessed at https://fred.stlouisfed.org.
Key Variables:
The four key variables in this project are real interest rate and real GDP. 
Real interest rate: A real interest rate is an interest rate that has been adjusted to remove the effects of inflation to reflect the real cost of funds to the borrower and the real yield to the lender or to an investor. The real interest rate reflects the rate of time-preference for current goods over future goods. The real interest rate of an investment is calculated as the difference between the nominal interest rate and the inflation rate:
Real Interest Rate = Nominal Interest Rate - Inflation (Expected or Actual)

Real GDP: Real gross domestic product (GDP) is an inflation-adjusted measure that reflects the value of all goods and services produced by an economy in a given year (expressed in base-year prices) and is often referred to as constant-price GDP, inflation-corrected GDP, or constant dollar GDP.